VertexBufferObject asserts isComplete()
This way it's still safe to create a VertexBufferObject
directly from a builder.  However, if it's not complete
the program will at least exit instead of continuing
on.

Note using VertexBufferObjectBuilder's toVertexBuffer()
method is still the preferred way to create the VBO.  In
fact, if the builder isn't complete, using
toVertexBuffer() instead of new VertexBufferObject will
throw a BasicException that can be caught and recovered
from.